Gravity with Mass
So in order to simulate gravity in space I made a script that can be attached to an empty game object. This script uses tags to find objects and according to their rigidbody mass they will move around each other based on the basic laws of newton. The scene I added so demo these features. But you'll have to edit the variables every time if you move anything. The mass of each rigidbody and the AddForce in the script Start.